backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
anthropic==0.21.3
python-docx==0.8.11
python-pptx==0.6.21
pydantic[email]==2.5.0
